id로 select, delete 하는 방식을 기본으로 제공하는 메소드로 변경

+++ src/main/java/com/b2en/sms/repo/ContChngHistRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.b2en.sms.entity.ContChngHist;
 
-public interface ContChngHistRepository extends JpaRepository<ContChngHist, Long>{
+public interface ContChngHistRepository extends JpaRepository<ContChngHist, Integer>{
 
 }
